fix(sdk): Move local user to local
