+++ python/src/numerous/frameworks/fastapi.py
@@ -5,7 +5,7 @@ from typing import Dict
 from fastapi import Request
 
 from numerous import user_session
-from numerous.utils import is_local_mode
+from numerous.local import is_local_mode, local_user
 
 
 class FastapiCookieGetter:
@@ -16,7 +16,7 @@ class FastapiCookieGetter:
         """Get the cookies associated with the current request."""
         if is_local_mode():
             # Update the cookies on the fastapi server
-            self.request.cookies.update(user_session.get_encoded_user_info())
+            self.request.cookies.update(user_session.get_encoded_user_info(local_user))
 
         return {str(key): str(val) for key, val in self.request.cookies.items()}
