+++ python/src/numerous/frameworks/flask.py
@@ -5,7 +5,7 @@ from typing import Dict
 from flask import request
 
 from numerous import user_session
-from numerous.utils import is_local_mode
+from numerous.local import is_local_mode, local_user
 
 
 class FlaskCookieGetter:
@@ -14,7 +14,7 @@ class FlaskCookieGetter:
         local_cookies = {}
         if is_local_mode():
             # Update the cookies on the flask server
-            local_cookies = user_session.get_encoded_user_info()
+            local_cookies = user_session.get_encoded_user_info(local_user)
         return {
             **local_cookies,
             **{str(key): str(val) for key, val in request.cookies.items()},
