+++ python/src/numerous/frameworks/panel.py
@@ -5,7 +5,7 @@ from typing import Dict
 import panel as pn
 
 from numerous import user_session
-from numerous.utils import is_local_mode
+from numerous.local import is_local_mode, local_user
 
 
 class PanelCookieGetter:
@@ -14,7 +14,7 @@ class PanelCookieGetter:
         """Get the cookies associated with the current request."""
         if is_local_mode():
             # Add the user info to the cookies on panel server
-            pn.state.cookies.update(user_session.get_encoded_user_info())
+            pn.state.cookies.update(user_session.get_encoded_user_info(local_user))
 
         if pn.state.curdoc and pn.state.curdoc.session_context:
             return {key: str(val) for key, val in pn.state.cookies.items()}
