+++ python/src/numerous/local.py
@@ -1,0 +1,13 @@
+"""Local mode utilities."""
+
+from os import getenv
+
+from numerous.user import User
+
+
+local_user = User(id="local_user", name="Local User")
+
+
+def is_local_mode() -> bool:
+    url = getenv("NUMEROUS_API_URL")
+    return url is None
