+++ python/src/numerous/utils.py
@@ -1,6 +1,5 @@
 """Miscellaneous utilities."""
 
-from os import getenv
 from typing import TypeVar
 
 
@@ -11,8 +10,3 @@ class _MissingType:
 MISSING = _MissingType()
 
 AppT = TypeVar("AppT")
-
-
-def is_local_mode() -> bool:
-    url = getenv("NUMEROUS_API_URL")
-    return url is None
